Revert some of the standalone code so that this PR doesn't have so many changes
