--- cli/denort/mod.rs
@@ -1,3 +1,0 @@
-// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-
-pub mod code_cache;
